refactor: DTO 패키지 분리
