+++ src/main/java/cart/application/CartItemService.java
@@ -3,9 +3,9 @@ package cart.application;
 import cart.domain.CartItem;
 import cart.domain.Member;
 import cart.domain.Product;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
+import cart.dto.request.CartItemQuantityUpdateRequest;
+import cart.dto.request.CartItemRequest;
+import cart.dto.response.CartItemResponse;
 import cart.exception.CartItemException.NotFound;
 import cart.exception.ProductException;
 import cart.repository.CartItemRepository;
