+++ src/main/java/cart/dto/response/CartItemResponse.java
@@ -1,11 +1,12 @@
-package cart.dto;
+package cart.dto.response;
 
 import cart.domain.CartItem;
 
 public class CartItemResponse {
-    private Long id;
-    private int quantity;
-    private ProductResponse product;
+
+    private final Long id;
+    private final int quantity;
+    private final ProductResponse product;
 
     private CartItemResponse(Long id, int quantity, ProductResponse product) {
         this.id = id;
