+++ src/main/java/cart/dto/response/ProductResponse.java
@@ -1,12 +1,13 @@
-package cart.dto;
+package cart.dto.response;
 
 import cart.domain.Product;
 
 public class ProductResponse {
-    private Long id;
-    private String name;
-    private int price;
-    private String imageUrl;
+
+    private final Long id;
+    private final String name;
+    private final int price;
+    private final String imageUrl;
 
     private ProductResponse(Long id, String name, int price, String imageUrl) {
         this.id = id;
