+++ src/test/java/cart/ui/CartItemApiControllerTest.java
@@ -27,9 +27,9 @@ import cart.configuration.resolver.MemberArgumentResolver;
 import cart.domain.CartItem;
 import cart.domain.Member;
 import cart.domain.Product;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
+import cart.dto.request.CartItemQuantityUpdateRequest;
+import cart.dto.request.CartItemRequest;
+import cart.dto.response.CartItemResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import org.apache.http.HttpHeaders;
